BEAST extension
I wrote an extensions allowing to use HMatrices for the boundary element
toolkit BEAST.jl.
The extension adds BEAST.jl as a weak dependency which means that BEAST.jl is not
required as a packege but an optional add on. Only if BEAST.jl is loaded before
HMatrices the extension will be available.